Start creating on_startup method, add router
