+++ tgbot/handlers/__init__.py
@@ -1,1 +1,1 @@
-from . import callbacks, echo, errors, student, teacher  # noqa
+from . import echo  # noqa
